Moved vehicle interfaces
